Reuse code for graphicspath autocompletion
