+++ src/nl/hannahsten/texifyidea/completion/pathcompletion/LatexFolderProvider.kt
@@ -2,13 +2,13 @@ package nl.hannahsten.texifyidea.completion.pathcompletion
 
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiFile
-import nl.hannahsten.texifyidea.completion.pathcompletion.LatexPathProviderBase
 
 /**
  * @author Lukas Heiligenbrunner
  */
 class LatexFolderProvider : LatexPathProviderBase() {
     override fun selectScanRoots(file: PsiFile): ArrayList<VirtualFile> {
+        // todo import package (subimport esp.)
         return getProjectRoots()
     }
